create-badge-model-add-seeder-add-policy

+++ app/Http/Controllers/TodoController.php
@@ -22,9 +22,8 @@ class TodoController extends Controller
     }
 
     private function sendReminderEmailTodo(Todo $todo,$id) {
-       
         $data = [$todo,$id];
-         $checkDate = empty($todo['date']) ? "date is empty" : "date not empty";
+        $checkDate = empty($todo['date']) ? "date is empty" : "date not empty";
          Storage::disk('local')->append('error_log/error_check.txt', $todo['date']." <-date| reminder-> ".$todo['toggle_reminder'].'| empty?: '.$checkDate.'| carbon date:'.Carbon::parse($todo['date'],'Asia/Kuala_Lumpur')->toString().' | date now:'.Carbon::now('Asia/Kuala_Lumpur'));
         $date = $todo['date'];
         if($todo['toggle_reminder'] == 1 && empty($date) == false) {
@@ -47,13 +46,11 @@ class TodoController extends Controller
             'date' => 'nullable|string|max:255',
         ]);
 
-    
         $request['toggle_reminder'] = !empty($request['date']) ? true : false;
-
+        /** @var \App\Models\user */
         $user = Auth::user();
         
-    
-         $todo = Todo::create([
+        $todo = Todo::create([
             'user_id'=> $user->id,
             'title' => $validatedTodo['title'],
             'description' => $validatedTodo['description'],
@@ -61,14 +58,8 @@ class TodoController extends Controller
             'toggle_reminder' => $request['toggle_reminder'],
         ]);
 
-        /** @var \App\Models\user */
-        $user = Auth::user();
-        $milestonesArray = $this->milestones($user);
-        $count = $user['count'];
-        $sum = 1 + $count;
-        $user['count'] = $sum;
-        $user->save();
-             
+        $milestonesArray = $user->handleMilestones();
+    
         event(new EventTodoLog( $validatedTodo['title'],"Store New Todo item"));
         $this->sendReminderEmailTodo($todo,$user->id);
 
@@ -76,47 +67,13 @@ class TodoController extends Controller
             'message' => 'Sucessfully added One to-do list',
             'message_status' => 'SUCCESS',
             'to' => 'display',
-            'count' => $sum,
-            'milestones' =>[
-                'badge_1' => $milestonesArray['badge_1'],
-                'badge_2' => $milestonesArray['badge_2'],
-                'badge_3' => $milestonesArray['badge_3'],
-                'achievements' => $milestonesArray['achievements']
-            ],
-           
-         
+            'milestones' => $milestonesArray,
+            // below is milesstones data for reference
+            // 'notficition_status' => $noti_status,
+            // 'achievements' => $achievementCount,
+            // 'badge_id' => $badge->id,
+            // 'badge_name' => $badge->name,
         ]);
-    }
-
-    private function milestones($_user){
-        $badge_1 = $_user['badge_1'];
-        $badge_2 = $_user['badge_2'];
-        $badge_3 = $_user['badge_3'];
-        $number = ($_user['count']+1)/10;
-
-        if(is_int($number)){
-            $_user['achievements'] = $number;
-        }
-        
-        // none,pass,achieved
-        $pass = "pass";
-        if($_user['achievements'] == 2 && $badge_1 == "none"){
-            $_user['badge_1'] = $pass;
-        }
-        else if($_user['achievements'] == 5 && $badge_2 == "none"){
-            $_user['badge_2'] = $pass;
-        }
-        else if($_user['achievements'] == 10 && $badge_3 == "none"){
-            $_user['badge_3'] = $pass;
-        }
-        $_user->save();
-        
-        return [
-            'badge_1' => $_user['badge_1'],
-            'badge_2' => $_user['badge_2'],
-            'badge_3' => $_user['badge_3'],
-            'achievements' => $_user['achievements']
-        ];
     }
 
     /**
@@ -128,6 +85,18 @@ class TodoController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $todo = Todo::find($id);
+                /** @var \App\Models\user */
+                $user = Auth::user();
+        if ($user->cannot('update', $todo)) {
+           
+            return response()->json([
+                'message' => 'Other user cannot update other user todo list',
+                'message_status' => 'DENIED',
+                'to' => 'display'
+            ],403);
+        }
+
         $validatedTodo = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string|max:255',
@@ -143,7 +112,7 @@ class TodoController extends Controller
             $request['toggle_reminder'] = 1;  
         }
 
-        $todo = Todo::find($id);
+        
         $todo['title'] = $validatedTodo['title'];
         $todo['description'] = $validatedTodo['description'];
         $todo['date'] = $validatedTodo['date'];
@@ -163,12 +132,22 @@ class TodoController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-
      * @return \Illuminate\Http\Response
      */
     public function destroy($id) //todo $todo
     {
+        // if ($request->user()->cannot('update', $id)) {
+        /** @var \App\Models\user */
+        $user = Auth::user();
         $todo = Todo::find($id);
+        if ($user->cannot('delete', $todo)) {
+            return response()->json([
+                'message' => 'Other user cannot delete other user todo list',
+                'message_status' => 'DENIED',
+                'to' => 'display'
+            ],403);
+        }
+      
         $todo->delete();
 
         event(new EventTodoLog( $todo['title'],"Deleted a Todo item"));
